made changes to controller and entity

+++ src/main/java/com/koowakchai/store/service/TShoppingCartService.java
@@ -1,5 +1,6 @@
 package com.koowakchai.store.service;
 
 public interface TShoppingCartService {
-    public void saveOrUpdateCart(int userId, int subtypeId, int productId, int quantity) throws Exception;
+    public void saveOrUpdateCart(long userId, int subtypeId, int productId, int quantity) throws Exception;
+//    public void getCartItem(int)
 }
