+++ src/main/java/com/koowakchai/store/service/impl/TShoppingCartServiceImpl.java
@@ -1,7 +1,7 @@
 package com.koowakchai.store.service.impl;
 
+import com.koowakchai.hibernate.entity.TShoppingCartEntity;
 import com.koowakchai.store.dao.TShoppingCartDao;
-import com.koowakchai.store.entity.TShoppingCartEntity;
 import com.koowakchai.store.service.TShoppingCartService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,7 +12,7 @@ public class TShoppingCartServiceImpl implements TShoppingCartService {
     @Autowired
     private TShoppingCartDao tShoppingCartDao;
 
-    public void saveOrUpdateCart(int userId, int subtypeId, int productId, int quantity) throws Exception{
+    public void saveOrUpdateCart(long userId, int subtypeId, int productId, int quantity) throws Exception{
         TShoppingCartEntity tShoppingCartEntity = new TShoppingCartEntity();
         tShoppingCartEntity.setProductId(productId);
         tShoppingCartEntity.setQuantity(quantity);
